requirements.txt
pyinstaller
requests
pygame
pillow
customtkinter
